Polish backtest changes and tests for trend gating and 2R TP
Agent-Logs-Url: https://github.com/Alaashoky/GOLDAPI/sessions/4f7cce09-5bb5-41f8-a874-d69856b77486

+++ src/goldbot/app/cli.py
@@ -211,7 +211,7 @@ def main() -> None:
         print(
             "Summary: PF={pf:.3f} | Trades={trades:.0f} | WinRate={wr:.2%} | PnL={pnl:.2f} | MaxDD={dd:.2f}".format(
                 pf=float(metrics.get("profit_factor", 0.0)),
-                trades=float(metrics.get("trades", 0.0)),
+                trades=int(metrics.get("trades", 0.0)),
                 wr=float(metrics.get("win_rate", 0.0)),
                 pnl=float(metrics.get("total_pnl", 0.0)),
                 dd=float(metrics.get("max_drawdown", 0.0)),
